git-svn-id: https://svn.code.sf.net/p/highrisedev/code@104 f146f8d7-247d-460d-a1b3-d0a612191574

+++ AI/floorAgent.cpp
@@ -40,8 +40,8 @@ bool FloorAgent::AddFloorSpace (FloorBase* pFS, int x, int x2 )
 {
    if (mLevel->TestForEmptySpace (x, x2))
    {
-      pFS->SetX (x);
-      pFS->SetX2 (x2);
+      pFS->SetX ((float)x);
+      pFS->SetX2 ((float)x2);
       mLevel->AddFloor (pFS);
    }
    return true;
